+++ scripts/systems/generic/generic.js
@@ -1,0 +1,6 @@
+import EffectDefinitionsGeneric from './effect-definitions-generic.js';
+
+export default {
+  DEFAULT_EFFECT_DEFINITIONS: {},
+  DEFAULT_STATUS_EFFECT_NAMES: [],
+};
