+++ scripts/systems/sw5e/sw5e.js
@@ -1,0 +1,6 @@
+import EffectDefinitionsSw5e from './effect-definitions-sw5e.js';
+
+export default {
+  DEFAULT_EFFECT_DEFINITIONS: {},
+  DEFAULT_STATUS_EFFECT_NAMES: [],
+};
